* smtpserver/smtpserver.c, smtpserver/policytest.h, . smtpserver/mxverify.c,   smtpserver/policytest.c, . smtpserver/cfgread.c,    man/smtpserver.8.in: .   - PARAM policydb-submit  -- separate DB for SUBMISSION .   - GCC-4.0.0 fixes.

+++ smtpserver/mxverify.c
@@ -249,7 +249,7 @@ dnsmxlookup(state, host, depth, mxmode, qtype)
 
 	if (debug)
 	  printf("000-  nscount=%d (== 0 ?)  arcount=%d  eom-cp=%d\n",
-		 nscount, arcount, eom-cp);
+		 nscount, arcount, (int)(eom-cp));
 
 	/* Ok, can continue to pick the ADDITIONAL SECTION data */
